Load Twilio settings from ENV, then from disk
Change the way Twilio settings are loaded into the app. The token strings are stored in ENV (put there through interaction with TwilioWrapper). If the ENV isn't set, it looks for the settings in config/twilio.json.

This works better for Heroku deploys, as well keeping things flexible for testing.